+++ src/common/utils.ts
@@ -81,8 +81,13 @@ export function isEqual(obj1, obj2) {
   return true;
 }
 
+// 很费性能，慎重用
 export const dataType = (val: any): string => {
   return Object.prototype.toString
     .call(val)
     .replace(/^.{8}(.+)]$/, (m, $1) => $1.toLowerCase());
+};
+
+export const mockInt = (minNum, maxNum) => {
+  return parseInt(Math.random() * (maxNum - minNum + 1) + minNum, 10);
 };